refactor(vector): refactor file and type names
